bug: deploy bug fix
