+++ src/app/page.tsx
@@ -1,17 +1,10 @@
-import TodayChatInfo from 'components/organisms/TodayChatInfo/index.tsx'
-import {
-  ChatInfoCard,
-  FilterModal,
-  Signup,
-  WorkerInfoCard,
-} from 'components/organisms/index.ts'
+import { FilterModal, TodayChatInfo } from 'components/organisms/index.ts'
 
 export default function Home() {
   return (
     <main>
-      {/* <TodayChatInfo />
-      <FilterModal /> */}
-      <WorkerInfoCard />
+      <TodayChatInfo />
+      <FilterModal />
     </main>
   )
 }
