+++ src/components/organisms/index.ts
@@ -3,3 +3,5 @@ export { default as FilterModal } from './FilterModal/index.tsx'
 export { default as Login } from './Login/index.tsx'
 export { default as Signup } from './Signup/index.tsx'
 export { default as WorkerInfoCard } from './WorkerInfoCard/index.tsx'
+export { default as TodayChatInfo } from './TodayChatInfo/index.tsx'
+export { default as AlertContainer } from './AlertContainer/index.tsx'
